refactor(emit): relocate transform_generator_function out of oversized async_es5_ir.rs
Move the full function* ES5 transform (and its rest-parameter helpers) into the
dedicated async_es5_ir_generator_fn.rs module. async_es5_ir.rs drops from 5094 to
4918 lines, so this PR net-reduces the over-cap file while keeping all new
generator emit logic in under-cap shards (generator_fn.rs 188, for_of.rs 1479).
Pure code relocation, no behavior change (§19 hygiene, per review).
